+++ source/rtcp_util.c
@@ -11,7 +11,7 @@
 #include "rtp/rtcp_header.h"
 #include "rtp/rtcp_util.h"
 
-int rtcp_type(const uint8_t *buffer, int size)
+int rtcp_type(const uint8_t *buffer, size_t size)
 {
     assert(buffer != NULL);
